+++ src/Tag.js
@@ -3,8 +3,9 @@ import classNames from 'classnames';
 import './styles/Tag.scss';
 
 import icBEM from './utils/icBEM';
+import prefixClass from './utils/prefixClass';
 
-const COMPONENT_NAME = 'ic-tag';
+const COMPONENT_NAME = prefixClass('tag');
 const ROOT_BEM = icBEM(COMPONENT_NAME);
 
 function Tag({ className, children }) {