+++ src/utils/prefixState.js
@@ -1,0 +1,17 @@
+import prefixClass from './prefixClass';
+
+/**
+ * Prefix anything passed-in with 'gyp-state-'.
+ *
+ * @param {String} state
+ * @return {String} result
+ */
+function prefixState(state) {
+    if (!(typeof state === 'string') || !state.length) {
+        throw new Error('state should be a non-empty string.');
+    }
+
+    return prefixClass(`state-${state}`);
+}
+
+export default prefixState;